Add .defaultStringBased when deriving an enumeration
