+++ core/src/main/scala-2/sttp/tapir/macros/SchemaMacros.scala
@@ -1,10 +1,10 @@
 package sttp.tapir.macros
 
 import magnolia1.Magnolia
+import sttp.tapir.Schema
 import sttp.tapir.generic.Configuration
 import sttp.tapir.generic.auto.SchemaMagnoliaDerivation
 import sttp.tapir.internal.{ModifySchemaMacro, OneOfMacro, SchemaEnumerationMacro, SchemaMapMacro}
-import sttp.tapir.{Schema, SchemaAnnotations, SchemaType, Validator}
 
 trait SchemaMacros[T] {
 
@@ -82,32 +82,9 @@ trait SchemaCompanionMacros extends SchemaMagnoliaDerivation {
     macro SchemaEnumerationMacro.derivedEnumerationValueCustomise[T]
 
   /** Create a schema for an [[Enumeration]], where the validator is created using the enumeration's values. The low-level representation of
-    * the enum is a `String`, and the enum values are encoded (so that they can be displayed in the documentation) using `.toString`.
+    * the enum is a `String`, and the enum values in the documentation will be encoded using `.toString`.
     */
   implicit def derivedEnumerationValue[T <: scala.Enumeration#Value]: Schema[T] = macro SchemaEnumerationMacro.derivedEnumerationValue[T]
 }
 
-class CreateDerivedEnumerationSchema[T](validator: Validator.Enumeration[T], schemaAnnotations: SchemaAnnotations[T]) {
 
-  // needed for binary compatibility
-  def this(validator: Validator.Enumeration[T]) = this(validator, SchemaAnnotations.empty)
-
-  /** @param encode
-    *   Specify how values of this type can be encoded to a raw value (typically a [[String]]; the raw form should correspond with
-    *   `schemaType`). This encoding will be used when generating documentation.
-    * @param schemaType
-    *   The low-level representation of the enumeration. Defaults to a string.
-    */
-  def apply(
-      encode: Option[T => Any] = None,
-      schemaType: SchemaType[T] = SchemaType.SString[T](),
-      default: Option[T] = None
-  ): Schema[T] = {
-    val v = encode.fold(validator)(e => validator.encode(e))
-
-    val s0 = Schema(schemaType).validate(v)
-    val s1 = default.fold(s0)(d => s0.default(d, encode.map(e => e(d))))
-
-    schemaAnnotations.enrich(s1)
-  }
-}
